move back to logic
