+++ src/main/java/org/jabref/gui/copyfiles/CopyFilesTask.java
@@ -15,11 +15,11 @@ import java.util.function.BiFunction;
 import javafx.concurrent.Task;
 
 import org.jabref.logic.l10n.Localization;
+import org.jabref.logic.util.FileUtil;
 import org.jabref.logic.util.OS;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.LinkedFile;
-import org.jabref.model.util.FileUtil;
 import org.jabref.model.util.OptionalUtil;
 import org.jabref.preferences.PreferencesService;
