+++ src/main/java/org/jabref/gui/copyfiles/CopySingleFileAction.java
@@ -10,9 +10,9 @@ import org.jabref.gui.DialogService;
 import org.jabref.gui.actions.SimpleCommand;
 import org.jabref.gui.util.DirectoryDialogConfiguration;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.logic.util.FileUtil;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.LinkedFile;
-import org.jabref.model.util.FileUtil;
 import org.jabref.model.util.OptionalUtil;
 import org.jabref.preferences.FilePreferences;
