+++ src/main/java/org/jabref/gui/externalfiles/ExternalFilesEntryLinker.java
@@ -13,10 +13,10 @@ import org.jabref.logic.cleanup.MoveFilesCleanup;
 import org.jabref.logic.cleanup.RenamePdfCleanup;
 import org.jabref.logic.pdf.search.indexing.IndexingTaskManager;
 import org.jabref.logic.pdf.search.indexing.PdfIndexer;
+import org.jabref.logic.util.FileUtil;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.LinkedFile;
-import org.jabref.model.util.FileUtil;
 import org.jabref.preferences.FilePreferences;
 
 import org.slf4j.Logger;
