+++ src/main/java/org/jabref/gui/externalfiles/ImportHandler.java
@@ -33,6 +33,7 @@ import org.jabref.logic.importer.fetcher.ArXivFetcher;
 import org.jabref.logic.importer.fetcher.DoiFetcher;
 import org.jabref.logic.importer.fileformat.BibtexParser;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.logic.util.FileUtil;
 import org.jabref.logic.util.UpdateField;
 import org.jabref.model.FieldChange;
 import org.jabref.model.database.BibDatabaseContext;
@@ -43,7 +44,6 @@ import org.jabref.model.entry.identifier.DOI;
 import org.jabref.model.groups.GroupEntryChanger;
 import org.jabref.model.groups.GroupTreeNode;
 import org.jabref.model.util.FileUpdateMonitor;
-import org.jabref.model.util.FileUtil;
 import org.jabref.model.util.OptionalUtil;
 import org.jabref.preferences.PreferencesService;
