+++ src/main/java/org/jabref/gui/maintable/columns/LibraryColumn.java
@@ -5,7 +5,7 @@ import org.jabref.gui.maintable.MainTableColumnModel;
 import org.jabref.gui.maintable.MainTableColumnModel.Type;
 import org.jabref.gui.util.ValueTableCellFactory;
 import org.jabref.logic.l10n.Localization;
-import org.jabref.model.util.FileUtil;
+import org.jabref.logic.util.FileUtil;
 
 public class LibraryColumn extends MainTableColumn<String> {
