+++ src/main/java/org/jabref/gui/openoffice/DetectOpenOfficeInstallation.java
@@ -11,9 +11,9 @@ import org.jabref.gui.desktop.os.NativeDesktop;
 import org.jabref.gui.util.DirectoryDialogConfiguration;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.openoffice.OpenOfficePreferences;
+import org.jabref.logic.util.FileUtil;
 import org.jabref.logic.util.OS;
 import org.jabref.model.strings.StringUtil;
-import org.jabref.model.util.FileUtil;
 
 /**
  * Tools for automatically detecting OpenOffice or LibreOffice installations.
