+++ src/main/java/org/jabref/gui/preferences/customimporter/CustomImporterTabViewModel.java
@@ -18,8 +18,8 @@ import org.jabref.gui.preferences.PreferenceTabViewModel;
 import org.jabref.gui.util.FileDialogConfiguration;
 import org.jabref.logic.importer.fileformat.CustomImporter;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.logic.util.FileUtil;
 import org.jabref.logic.util.StandardFileType;
-import org.jabref.model.util.FileUtil;
 import org.jabref.preferences.PreferencesService;
 
 import org.slf4j.Logger;
