+++ src/main/java/org/jabref/logic/git/GitHandler.java
@@ -7,7 +7,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Optional;
 
-import org.jabref.model.util.FileUtil;
+import org.jabref.logic.util.FileUtil;
 
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.RmCommand;
