+++ src/main/java/org/jabref/logic/importer/fileformat/PdfGrobidImporter.java
@@ -11,10 +11,10 @@ import org.jabref.logic.importer.ImportFormatPreferences;
 import org.jabref.logic.importer.Importer;
 import org.jabref.logic.importer.ParserResult;
 import org.jabref.logic.importer.util.GrobidService;
+import org.jabref.logic.util.FileUtil;
 import org.jabref.logic.util.StandardFileType;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.LinkedFile;
-import org.jabref.model.util.FileUtil;
 
 /**
  * Wraps the GrobidService function to be used as an Importer.
