+++ src/main/java/org/jabref/logic/util/io/BackupFileUtil.java
@@ -10,8 +10,8 @@ import java.util.HexFormat;
 import java.util.Optional;
 
 import org.jabref.logic.util.BackupFileType;
+import org.jabref.logic.util.FileUtil;
 import org.jabref.logic.util.OS;
-import org.jabref.model.util.FileUtil;
 
 import net.harawata.appdirs.AppDirsFactory;
 import org.slf4j.Logger;
