+++ src/main/java/org/jabref/logic/util/io/FileNameCleaner.java
@@ -1,4 +1,6 @@
-package org.jabref.model.util;
+package org.jabref.logic.util.io;
+
+import org.jabref.logic.util.FileUtil;
 
 /**
  * This class is based on http://stackoverflow.com/a/5626340/873282
