+++ src/main/java/org/jabref/logic/util/io/FileNameUniqueness.java
@@ -10,7 +10,7 @@ import java.util.regex.Pattern;
 
 import org.jabref.gui.DialogService;
 import org.jabref.logic.l10n.Localization;
-import org.jabref.model.util.FileUtil;
+import org.jabref.logic.util.FileUtil;
 
 public class FileNameUniqueness {
     private static final Pattern DUPLICATE_MARK_PATTERN = Pattern.compile("(.*) \\(\\d+\\)");
