+++ src/main/java/org/jabref/logic/util/io/RegExpBasedFileFinder.java
@@ -20,7 +20,6 @@ import java.util.stream.Stream;
 import org.jabref.logic.citationkeypattern.BracketedPattern;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.strings.StringUtil;
-import org.jabref.model.util.FileNameCleaner;
 
 class RegExpBasedFileFinder implements FileFinder {
