+++ src/main/java/org/jabref/model/groups/TexGroup.java
@@ -11,11 +11,11 @@ import java.util.Set;
 import org.jabref.architecture.AllowedToUseLogic;
 import org.jabref.logic.auxparser.AuxParser;
 import org.jabref.logic.auxparser.AuxParserResult;
+import org.jabref.logic.util.FileUtil;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.metadata.MetaData;
 import org.jabref.model.util.FileUpdateListener;
 import org.jabref.model.util.FileUpdateMonitor;
-import org.jabref.model.util.FileUtil;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
