+++ src/test/java/org/jabref/logic/crawler/CrawlerTest.java
@@ -13,10 +13,10 @@ import org.jabref.logic.exporter.SavePreferences;
 import org.jabref.logic.git.SlrGitHandler;
 import org.jabref.logic.importer.ImportFormatPreferences;
 import org.jabref.logic.importer.ImporterPreferences;
+import org.jabref.logic.util.FileUtil;
 import org.jabref.model.entry.BibEntryTypesManager;
 import org.jabref.model.metadata.SaveOrderConfig;
 import org.jabref.model.util.DummyFileUpdateMonitor;
-import org.jabref.model.util.FileUtil;
 import org.jabref.preferences.GeneralPreferences;
 import org.jabref.testutils.category.FetcherTest;
