+++ src/test/java/org/jabref/logic/crawler/StudyYamlParserTest.java
@@ -4,10 +4,10 @@ import java.net.URL;
 import java.nio.file.Path;
 import java.util.List;
 
+import org.jabref.logic.util.FileUtil;
 import org.jabref.model.study.Study;
 import org.jabref.model.study.StudyDatabase;
 import org.jabref.model.study.StudyQuery;
-import org.jabref.model.util.FileUtil;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
